cleaning up the database interface and added more tests.
there are some strange uses of Lock() and Unlock() that have
been commented; they should be removed, or fixed to be properly
locking things. locks should never be exposed outside the data
structure that it is locking (that's asking for deadlock trouble).

made tests not depend on the internals of the database marshalling
(gob is used, but the tests should not know this, e.g. if we
later decide to rewrite the internal storage)

the Has() function uses the Get() function instead of using a
linear order O(N) search through the DB.

i've added a few TODOs that needs to be answered by Heine.