added Equal() and String() and new type codeReviewID.
To make things more general, I've added GobEncode and GobDecode
to allow external stuff use those for more general types,
such as codeReviewID (which is an int) to be saved to the DB.
It feels a bit strange, but it works ok.
Added also a few more tests, including one with concurrent
NewCodeReview() calls. Deleted one test, which exposed
the internals of codereview, and renamed to unexported
function nextCodeReviewID() instead. Thus, now we need to
create a new code review to learn what is the next code review id.