entities/members: removed in-memory cache and complicated locking.
This commit removes the in-memory map-based cache used to store
members. This implementation involved complicated locking mechanisms
which is likely to cause deadlock problems. Moreover, BoltDB is
already a memory backed database (Memory-Maped File), so the
performance gain was not very high (about 40 %). Probably the main
reason for better performance with map-based cache is that we can
then avoid marshalling/unmarshalling whenever there is a cache hit.

Will look into better ways to improve performance instead.
For example, I want to understand why JSON is faster than GOB,
which does not make sense.