ci-refactor: removed TODO
After some additional attempts at trying to use the channel
abstraction instead of condition variables to solve the
fine-grained locking issue, I've not come up with anything
better. But the condition variables are working fine, and
it even passes the -race detector.

The old version can be tested on this commit by removing the x
in front of xTestConcurrentLocking2. By running go test -v
you will be able to see that it deadlocks. Also when this
test is enabled, the race detector never finishes (before the
deadline of 10min). Hence, it is safe to assume that the old
version in _lockold.go is flawed.

I will remove it after this commit and commit again; just to
keep the history.