Hack to fix group repo creation
Sometimes creating a group fails because the hook was already created;
probably because the group wasn't fully created in a previous attempt
(e.g. because it didn't finish) thus trying again should not fail
the rest of the process.

BUT: This maybe needs more careful consideration in the future
because we shouldn't really allow half-way creation of things on
github. How can we do multiple API calls against github in a
transaction? (and if the whole set of API calls fails, how can
we roll back?)